requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
anywidget==0.9.18
appdirs==1.4.4
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
contourpy==1.3.2
crashtest==0.4.1
cryptography==44.0.2
cycler==0.12.1
Cython==3.1.1
dataclasses-json==0.6.7
datasets==3.5.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
emoji==2.14.1
eval_type_backport==0.2.2
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
fonttools==4.58.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
google-genai==1.12.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
hdbscan==0.8.40
html5lib==1.1
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
installer==0.7.0
ipywidgets==8.1.7
isort==6.0.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jq==1.8.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.15
keyring==25.6.0
kiwisolver==1.4.8
langchain==0.3.24
langchain-anthropic==0.3.10
langchain-community==0.3.23
langchain-core==0.3.56
langchain-google-genai==2.1.2
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.3.29
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.1.147
llvmlite==0.44.0
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.35.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
olefile==0.47
openai==1.74.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pbs-installer==2025.4.9
pillow==11.2.1
pkginfo==1.12.1.2
plotly==6.1.2
poetry-core==2.1.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
psygnal==0.13.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
pynndescent==0.5.13
pyparsing==3.2.3
pypdf==5.4.0
pyproject_hooks==1.2.0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
ragas==0.2.15
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.4.2
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==4.0.2
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
streamlit==1.44.1
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenize_rt==6.2.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
trove-classifiers==2025.4.11.15
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
unstructured==0.17.2
unstructured-client==0.34.0
urllib3==2.4.0
uvicorn==0.34.1
virtualenv==20.30.0
webencodings==0.5.1
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xattr==1.1.4
xxhash==3.5.0
yarl==1.19.0
zstandard==0.23.0
